requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Limiter>=3.5.0
gunicorn>=21.0.0
psycopg2-binary
openpyxl>=3.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
